requirements.txt
# requirements.txt (base - works everywhere)
keras
numpy
pillow
scikit-learn
pyyaml
matplotlib
seaborn
mlflow
tensorflow-datasets